GIR Scraper: account for two-digit API versions
The scraper generator for GIRs was only checking the first digit of
versions when deciding whether to fallback to the namespace version of
a GIR or the largest API change (ie. member annotate with `Since`).

For example, `10 API` will now be chosen over `1.24+`, where previous
the latter would be preferred.