icons: override get_items() in sprites.thor to return gnome icons
In upstream DevDocs the static icons SCSS file was replaced with a
Ruby template that automagically computes icon offsets in the sprites.

Remove our statically generated icon sprites and replace them with a
JSON file that maps docs to icon names. Then override the `get_icons()`
method in `sprites.thor` to return the proper icons.

This results in a lot of redundant icon sprites, but requires less
manual work and makes all the other icon related styles function.